+++ src/listener/worker.rs
@@ -118,35 +118,32 @@ where
     /// Returns only the messages, while the None returned by poll are discarded
     #[allow(clippy::needless_collect)]
     fn poll(&mut self) -> Result<(), mpsc::SendError<ListenerMsg<U>>> {
-        let msg: Vec<ListenerMsg<U>> = self
-            .ports
-            .iter_mut()
-            .filter_map(|x| {
-                if x.should_poll() {
-                    let msg = match x.poll() {
-                        Ok(Some(ev)) => Some(ListenerMsg::User(ev)),
-                        Ok(None) => None,
-                        Err(err) => Some(ListenerMsg::Error(err)),
-                    };
-                    // Update next poll
-                    x.calc_next_poll();
-                    msg
-                } else {
-                    None
+        let port_iter = self.ports.iter_mut().filter(|port| port.should_poll());
+
+        for port in port_iter {
+            let mut times_remaining = port.max_poll();
+            // poll a port until it has nothing anymore
+            loop {
+                let msg = match port.poll() {
+                    Ok(Some(ev)) => ListenerMsg::User(ev),
+                    Ok(None) => break,
+                    Err(err) => ListenerMsg::Error(err),
+                };
+
+                self.sender.send(msg)?;
+
+                // do this at the end to at least call it once
+                times_remaining = times_remaining.saturating_sub(1);
+
+                if times_remaining == 0 {
+                    break;
                 }
-            })
-            .collect();
-        // Send messages
-        match msg
-            .into_iter()
-            .map(|x| self.sender.send(x))
-            .filter(|x| x.is_err())
-            .map(|x| x.err().unwrap())
-            .next()
-        {
-            None => Ok(()),
-            Some(e) => Err(e),
+            }
+            // Update next poll
+            port.calc_next_poll();
         }
+
+        Ok(())
     }
 
     /// thread run method
@@ -185,6 +182,30 @@ mod test {
     use crate::core::event::{Key, KeyEvent};
     use crate::mock::{MockEvent, MockPoll};
     use crate::Event;
+
+    #[test]
+    fn worker_should_poll_multiple_times() {
+        let (tx, rx) = mpsc::channel();
+        let paused = Arc::new(RwLock::new(false));
+        let paused_t = Arc::clone(&paused);
+        let running = Arc::new(RwLock::new(true));
+        let running_t = Arc::clone(&running);
+
+        let mut mock_port = Port::new(Box::new(MockPoll::default()), Duration::from_secs(5));
+        mock_port.set_max_poll(10);
+
+        let mut worker =
+            EventListenerWorker::<MockEvent>::new(vec![mock_port], tx, paused_t, running_t, None);
+        assert!(worker.poll().is_ok());
+        assert!(worker.next_event() <= Duration::from_secs(5));
+        let mut recieved = Vec::new();
+
+        while let Ok(msg) = rx.try_recv() {
+            recieved.push(msg);
+        }
+
+        assert_eq!(recieved.len(), 10);
+    }
 
     #[test]
     fn worker_should_send_poll() {